requirements.txt
langchain-community
langchain-core
langchain-openai
PyPDF2
langchain
streamlit
python-dotenv
ipykernel